chore: replace `useModalModelDownload` with `ModelDownloadModal`
