+++ webui/react/src/components/ModelDownloadModal.module.scss
@@ -1,17 +1,23 @@
 .base {
-  .topLine {
+  .commandContainer {
+    align-items: center;
     display: flex;
+    gap: 12px;
     justify-content: space-between;
   }
   .codeSample {
     background-color: var(--theme-float-strong);
     border: solid var(--theme-stroke-width) var(--theme-float-border);
     border-radius: 4px;
+    display: inline-block;
     font-family: var(--theme-font-family-code);
-    margin-bottom: 8px;
+    overflow-x: scroll;
     padding: 8px;
+    white-space: nowrap;
+    width: 100%;
   }
   .bottomLine {
     color: var(--theme-float-on-weak);
+    margin-top: 8px;
   }
 }
