+++ webui/react/src/pages/ModelVersionDetails/ModelVersionHeader.tsx
@@ -6,11 +6,12 @@ import React, { useCallback, useMemo, useState } from 'react';
 import InfoBox, { InfoRow } from 'components/InfoBox';
 import Breadcrumb from 'components/kit/Breadcrumb';
 import Button from 'components/kit/Button';
+import { useModal } from 'components/kit/Modal';
 import Tags, { tagsActionHelper } from 'components/kit/Tags';
 import Avatar from 'components/kit/UserAvatar';
 import Link from 'components/Link';
+import ModelDownloadModal from 'components/ModelDownloadModal';
 import TimeAgo from 'components/TimeAgo';
-import useModalModelDownload from 'hooks/useModal/Model/useModalModelDownload';
 import useModalModelVersionDelete from 'hooks/useModal/Model/useModalModelVersionDelete';
 import useModalModelVersionEdit from 'hooks/useModal/Model/useModalModelVersionEdit';
 import usePermissions from 'hooks/usePermissions';
@@ -54,18 +55,13 @@ const ModelVersionHeader: React.FC<Props> = ({
   const users = Loadable.map(loadableUsers, ({ users }) => users);
   const [showUseInNotebook, setShowUseInNotebook] = useState(false);
 
-  const { contextHolder: modalModelDownloadContextHolder, modalOpen: openModelDownload } =
-    useModalModelDownload();
+  const modelDownloadModal = useModal(ModelDownloadModal);
 
   const { contextHolder: modalModelVersionDeleteContextHolder, modalOpen: openModalVersionDelete } =
     useModalModelVersionDelete();
 
   const { contextHolder: modalModelNameEditContextHolder, modalOpen: openModelNameEdit } =
     useModalModelVersionEdit({ fetchModelVersion, modelVersion });
-
-  const handleDownloadModel = useCallback(() => {
-    openModelDownload(modelVersion);
-  }, [modelVersion, openModelDownload]);
 
   const { canDeleteModelVersion, canModifyModelVersion } = usePermissions();
 
@@ -133,7 +129,7 @@ const ModelVersionHeader: React.FC<Props> = ({
         danger: false,
         disabled: false,
         key: 'download-model',
-        onClick: handleDownloadModel,
+        onClick: () => modelDownloadModal.open(),
         text: 'Download',
       },
       {
@@ -165,8 +161,8 @@ const ModelVersionHeader: React.FC<Props> = ({
     modelVersion,
     canModifyModelVersion,
     openModelNameEdit,
-    handleDownloadModel,
     canDeleteModelVersion,
+    modelDownloadModal,
     handleDelete,
   ]);
 
@@ -276,7 +272,7 @@ with det.import_from_path(path + "/code"):
         </div>
         <InfoBox rows={infoRows} separator={false} />
       </div>
-      {modalModelDownloadContextHolder}
+      <modelDownloadModal.Component version={modelVersion} />
       {modalModelVersionDeleteContextHolder}
       {modalModelNameEditContextHolder}
       <Modal
